Draft up requirements based on latest conversation
Here is a draft in progress for documenting what the purpose of the software is and what we expect it to do (the "requirements"). It is still a work in progress. Things in here so far are based on our most recent email threads as well as some ideas taken from my private notes.

This was drafted up as a way of getting the project started and as a starting point for more discussion. Feedback is welcome!